updated to new variable transformations. seems to work
