+++ test/runtests.jl
@@ -1,3 +1,10 @@
+##################################################################################
+# This file is part of StateSpaceEcon.jl
+# BSD 3-Clause License
+# Copyright (c) 2020, Bank of Canada
+# All rights reserved.
+##################################################################################
+
 using StateSpaceEcon
 using TimeSeriesEcon
 using ModelBaseEcon
